% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_experiment.R
\name{extract_experiment}
\alias{extract_experiment}
\title{Extract EXPERIMENT table from Labkey}
\usage{
extract_experiment(rodbc_chan, cultures = NULL, protocols = NULL,
  annees = NULL, essais = NULL)
}
\arguments{
\item{rodbc_chan}{channel odbc created by labkey_connect{RLabkey}}

\item{cultures}{a vector of character (length 1 = code culture)}

\item{protocols}{a vector of character (length 3 = protocol code)}

\item{annees}{a vector of years}
}
\value{
a experiment dataframe
}
\description{
Extract EXPERIMENT table from Labkey
}
\examples{
lbk <- labkey_connect(login="consult", mdp="cetiom")
experiment <- extract_experiment(lbk, cultures=c("T","C"), protocols="VCE", annees=c(2017,2018), essais=c("VCE31","VCE40"))

}
