\name{labkey.experiment.SAMPLE_DERIVATION_PROTOCOL}
\alias{labkey.experiment.SAMPLE_DERIVATION_PROTOCOL}
\title{Constant for the Simple Derivation Protocol}
\description{
Simple Derivation Protocol constant.
}
\details{
This value can be used in the labkey.experiment.saveBatch function when creating runs that aren't backed by an assay protocol.
}

\author{Karl Lum}
\seealso{
\code{\link{labkey.experiment.saveBatch}}
}
\keyword{IO}
