\name{labkey.rstudio.isInitialized}
\alias{labkey.rstudio.isInitialized}
\title{Check valid rlabkey session}
\description{
LabKey-RStudio integration helper. Not intended for use outside RStudio.
}
\usage{
labkey.rstudio.isInitialized()
}
\examples{
\donttest{
## RStudio console only
library(Rlabkey)
labkey.rstudio.isInitialized()

}
}
