\name{labkey.setDefaults}
\alias{labkey.setDefaults}
\title{Set the credentials used for all http or https requests}
\description{
Use this function to set the "apiKey" package environment variable to be used for all http or https requests.
You can also use labkey.setDefaults() without any parameters to reset/clear the settings.}
\arguments{
  \item{apiKey}{session key from your server}
}
