% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labkey_connect.R
\name{labkey_connect}
\alias{labkey_connect}
\title{labkey_connect}
\usage{
labkey_connect(login = "consult", mdp = "cetiom", dsn = "LK_CETIOM")
}
\arguments{
\item{login}{a character}

\item{mdp}{a character}

\item{dsn}{a character}
}
\value{
a RODBC
}
\description{
labkey_connect
}
\examples{
labkey_connect(login="consult", mdp="cetiom")
}
