% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labkey.R
\name{extract_labkey}
\alias{extract_labkey}
\title{Extract data from Labkey}
\usage{
extract_labkey(rodbc_chan, cultures = NULL, protocols = NULL,
  annees = NULL)
}
\arguments{
\item{rodbc_chan}{channel odbc created by labkey_connect{RLabkey}}

\item{cultures}{a vector of character (length 1 = code culture)}

\item{protocols}{a vector of character (length 3 = protocol code)}

\item{annees}{a vector of years}
}
\value{
a dataframe
}
\description{
Extract data from Labkey
}
\examples{
lbk <- labkey_connect(login="consult", mdp="cetiom")
data <- extract_labkey(lbk, cultures=c("T","C"), protocols="VCE", annees=c(2017,2018))

}
