\name{Rlabkey-package}
\alias{Rlabkey-package}
\alias{Rlabkey}
\docType{package}
\title{
Import data from a labkey data base into an R data frame
}
\description{
This package allows the import of data from a labkey data base
into an R data frame. The import can be accomplished with Sql
commands or by specifying the query schema information.
}
\details{
\tabular{ll}{
Package: \tab Rlabkey\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2008-08-18\cr
License: \tab Apache 2.0\cr
LazyLoad: \tab yes\cr
}
Need to insert text on how to use the package in general.
}
\author{
Valerie Obenchain
}
\references{http://www.omegahat.org/RCurl/,
http://dssm.unipa.it/CRAN/web/packages/rjson/rjson.pdf,
https://www.labkey.org/project/home/begin.view}
\keyword{ package }
\seealso{
\code{\link{labkey.selectRows}}, \code{\link{labkey.executeSql}}, \code{\link{makeFilter}}
}
\examples{
}

