\name{RlabkeyUsersGuide}
\alias{RlabkeyUsersGuide}
\title{Open the Rlabkey Users Guide}
\description{ Brings up the Rlabkey Users Guide.}
\usage{
RlabkeyUsersGuide(view=TRUE) 
}
\arguments{
\item{view}{Leave as default TRUE}
}

\details{Brings up the Rlabkey Users Guide.}

\value{Path to the Users Guide pdf.
}
\author{Peter Hussey}
\examples{
\donttest{

# library(Rlabkey)
RlabkeyUsersGuide()

}
}
\keyword{file}
